decouple environmentSetter.py; add python3 support for databaseAgent
add path to python protos file; add python3 compatiable for itervalues(), iterkeys() and isAlive()

add python3 support for server class and subprocess return value

+++ sam/base/routingMorphic.py
@@ -119,7 +119,7 @@ class RoutingMorphic(object):
         for i in range(bitsLength):
             mask = mask + (0x1 << i)
 
-        return self._sc.int2Bytes(mask, bitsLength/8)
+        return self._sc.int2Bytes(mask, int(bitsLength/8))
 
     def encodeIdentifierForSFC(self, sfciID, vnfID):
         # sfci.sfciID, vnf.vnfID
